Add arm64 support to linux_syscall_support.h.
This CL adds arm64 support for linux syscall support.  The arm64
kernel depricates a number of syscalls, such as fork(), poll() and open(), so
we need to replicate the behaviour of these syscalls from newer related
syscalls, such as clone(), ppoll() and openat().

The code added here is based on that in glibc for aarch64, and Android bionic
for arm64, but currently hasn't been tested on either platform.  It does
enable breakpad to compile for arm64 though.

BUG=354405,335641
R=jacob.bramley@arm.com, mseaborn@chromium.org

Review URL: https://codereview.chromium.org/220933002

Patch from Ross McIlroy <rmcilroy@chromium.org>.